fixes name clashes caused by leading numbers #133
